backend/pyproject.toml
[project]
name = "betweenus-backend"
version = "0.1.0"
description = "BetweenUs API and async workers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.31.0",
  "pydantic-settings>=2.6.0",
  "sqlmodel>=0.0.22",
  "celery>=5.4.0",
  "redis>=5.2.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.20",
  "httpx>=0.27.2"
]

[dependency-groups]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q --cov=app --cov-report=term-missing"
